Add LC calculation
This is inferred rather than explicitly stated,
since the official ruling is:

> Assign the armor whatever LC seems appropriate, often
> based more on its look and feel than its actual statistics.

This is at odds with rules to increase LC based on the fraction of maximum DR.
Examples from Ultra-Tech line up with regular flexible body armour
lying in the ¼ to ⅙ Max DR range at LC3,
and Tacsuits lying in the ½ to ¼ Max DR range.

There's outliers, like Bioplas being technically over the ⅙ line
implying it should have a LC decrement,
and gloves, helmets and boots should have an LC increment,
but since bioplas being lower LC could be interpreted as rounding error
and I want this for making full suits this is sufficient.